Move ExecutionBlocker to a better location
