+++ comfy/graph_utils.py
@@ -84,11 +84,6 @@ class GraphBuilder:
         id = self.prefix + id
         del self.nodes[id]
 
-# Return this from a node and any users will be blocked with the given error message.
-class ExecutionBlocker:
-    def __init__(self, message):
-        self.message = message
-
 class Node:
     def __init__(self, id, class_type, inputs):
         self.id = id
